refactor(background): split index.ts into domain modules
Trim src/background/index.ts from 1756 to 142 lines by moving
message handlers, lifecycle logic, and listener bodies into
domain modules. The MV3 service worker still registers all
listeners synchronously at module top level so wake-up events
aren't missed.

- messages/{suspension,sessions,cloud,pro,recovery,exportReminder,
  dedup,backupHealth,devtools}.ts: one handler per slice of the
  MessageType union; messages/index.ts is the central dispatcher
- lifecycle/{hibernation,onInstalled,onStartup}.ts: install/startup
  handlers and the post-startup re-discard guard
- state.ts, badge.ts, alarms.ts, contextMenus.ts: shared helpers
- listeners.ts: listener bodies (registrations stay in index.ts)

No behavior change. The MessageType union is unchanged.
